fix: Separated dependencies in a separate file
